+++ cobras/bavarde/bot/bot.py
@@ -1,0 +1,8 @@
+'''Chat bot
+
+Copyright (c) 2019 Machine Zone, Inc. All rights reserved.
+'''
+
+
+def runBot():
+    print('running bot')